Update Post Submit
Added in some changes, on task one, i set it up with two options. i left the grouping version, and added a different version that is more about resorting the array, and outputting that.  I also realized we have some non printable characters so i set the comments to filter through and remove those.